+++ Changelog.md
@@ -5,6 +5,7 @@
 * Correct `==` method for DocxStructure.
 * Correctly handle and warn if docx file do not contain `docProps/app.xml`
 * Extract GradientColor parsing to Class
+* Correct warn if document is password protected
 
 ### Fixes
 #### `DocxParser`